Put source files of the project in separate subfolder (to avoid filenames pollution).
