+++ benchmarks/tests/main_gprof.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 
 #include "matrixutils.h"
-#include "munkres.h"
+#include "munkres-cpp/munkres.h"
