+++ benchmarks/tests/main_hayai.cpp
@@ -2,7 +2,7 @@
 #include <vector>
 
 #include "matrixutils.h"
-#include "munkres.h"
+#include "munkres-cpp/munkres.h"
