+++ benchmarks/tests/main_rdtsc.cpp
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "matrixutils.h"
-#include "munkres.h"
+#include "munkres-cpp/munkres.h"
