+++ src/munkres-cpp/adapters/matrix_qt.h
@@ -16,65 +16,40 @@
  *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA
  */
 
-#if !defined(_MUNKRES_ADAPTERS_STD_2D_VECTOR_H_)
-#define _MUNKRES_ADAPTERS_STD_2D_VECTOR_H_
+#if !defined(_MUNKRES_ADAPTERS_QT_MATRIX_H_)
+#define _MUNKRES_ADAPTERS_QT_MATRIX_H_
 
-#include "munkres.h"
-#include <vector>
-
+#include "munkres-cpp/matrix_base.h"
+#include <QGenericMatrix>
 
 namespace munkres
 {
 
-template<class T>
-class matrix_std_2d_vector : public matrix_base<T>
+template<class T, const int N, const int M>
+class matrix_qt : public matrix_base<T>, QGenericMatrix<N, M, T>
 {
     public:
-        matrix_std_2d_vector (std::vector<std::vector<T>> & data)
-            : data {data}
-        {
-        }
-
         const T & operator () (const size_t row, const size_t column) const noexcept override
         {
-            return data [row][column];
+            return QGenericMatrix<N, M, T>::operator () (row, column) ;
         };
 
         T & operator () (const size_t row, const size_t column) noexcept override
         {
-            return data [row][column];
+            return QGenericMatrix<N, M, T>::operator () (row, column) ;
         }
 
         size_t columns () const noexcept override
         {
-            size_t columns = data.size () ? data [0].size () : 0;
-            for (size_t i = 0; i < data.size(); ++i) {
-                columns = std::min (columns, data [i].size () );
-            }
-            return columns;
+            return M;
         }
 
         size_t rows () const noexcept override
         {
-            return data.size ();
+            return N;
         }
-
-        void resize (const size_t rows, const size_t columns, const T value = matrix_base<T>::zero) override
-        {
-            if (rows != this->rows () ) {
-                data.resize (rows);
-                if (columns != this->columns () ) {
-                    for (size_t i = 0; i < rows; ++i) {
-                        data [i].resize (columns, value);
-                    }
-                }
-            }
-        }
-
-    private:
-        std::vector<std::vector<T>> & data;
 };
 
 }// namespace munkres
 
-#endif /* !defined(_MUNKRES_ADAPTERS_STD_2D_VECTOR_H_) */
+#endif /* !defined(_MUNKRES_ADAPTERS_QT_MATRIX_H_) */
