+++ src/munkres-cpp/adapters/matrix_std_2d_vector.h
@@ -16,45 +16,65 @@
  *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA
  */
 
-#if !defined(_MUNKRES_ADAPTERS_OPENCV_MATRIX_H_)
-#define _MUNKRES_ADAPTERS_OPENCV_MATRIX_H_
+#if !defined(_MUNKRES_ADAPTERS_STD_2D_VECTOR_H_)
+#define _MUNKRES_ADAPTERS_STD_2D_VECTOR_H_
 
-#include "matrix_base.h"
-#include <cv.hpp>
+#include "munkres-cpp/munkres.h"
+#include <vector>
+
 
 namespace munkres
 {
 
 template<class T>
-class matrix_opencv : public matrix_base<T>, cv::Mat_<T>
+class matrix_std_2d_vector : public matrix_base<T>
 {
     public:
-        matrix_opencv (const size_t rows, const size_t columns)
-            : cv::Mat_<T>::Mat_ (rows, columns)
+        matrix_std_2d_vector (std::vector<std::vector<T>> & data)
+            : data {data}
         {
         }
 
         const T & operator () (const size_t row, const size_t column) const noexcept override
         {
-            return cv::Mat_<T>::operator () (row, column) ;
+            return data [row][column];
         };
 
         T & operator () (const size_t row, const size_t column) noexcept override
         {
-            return cv::Mat_<T>::operator () (row, column) ;
+            return data [row][column];
         }
 
         size_t columns () const noexcept override
         {
-            return cv::Mat_<T>::cols;
+            size_t columns = data.size () ? data [0].size () : 0;
+            for (size_t i = 0; i < data.size(); ++i) {
+                columns = std::min (columns, data [i].size () );
+            }
+            return columns;
         }
 
         size_t rows () const noexcept override
         {
-            return cv::Mat_<T>::rows;
+            return data.size ();
         }
+
+        void resize (const size_t rows, const size_t columns, const T value = matrix_base<T>::zero) override
+        {
+            if (rows != this->rows () ) {
+                data.resize (rows);
+                if (columns != this->columns () ) {
+                    for (size_t i = 0; i < rows; ++i) {
+                        data [i].resize (columns, value);
+                    }
+                }
+            }
+        }
+
+    private:
+        std::vector<std::vector<T>> & data;
 };
 
 }// namespace munkres
 
-#endif /* !defined(_MUNKRES_ADAPTERS_OPENCV_MATRIX_H_) */
+#endif /* !defined(_MUNKRES_ADAPTERS_STD_2D_VECTOR_H_) */
