+++ src/munkres-cpp/munkres.h
@@ -19,7 +19,7 @@
 #if !defined(_MUNKRES_H_)
 #define _MUNKRES_H_
 
-#include "matrix.h"
+#include "munkres-cpp/matrix.h"
 
 #include <list>
 #include <utility>
