+++ tests/adapters/matrix_std_2d_vector_test.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "adapters/matrix_std_2d_vector.h"
+#include "munkres-cpp/adapters/matrix_std_2d_vector.h"
